add multiple worlds

+++ RobotGame/Position.h
@@ -1,8 +1,8 @@
 //
 //  Position.h
-//  Robots
+//  RobotGame
 //
-//  Created by Diana Zmuda on 9/11/12.
+//  Created by Diana Zmuda on 9/17/12.
 //  Copyright (c) 2012 Diana Zmuda. All rights reserved.
 //
 
@@ -10,8 +10,8 @@
 
 @interface Position : NSObject <NSCopying>
 
-@property double x;
-@property double y;
+@property int x;
+@property int y;
 
 +(Position*)withX:(double)x andY:(double)y;
